pyproject.toml
[project]
name = "redneck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

packages = [
  { include = "redneck" },
  { include = "redplug" }
]

dependencies = [
    "coloredlogs>=15.0.1",
    "importlib>=1.0.4",
    "pathlib>=1.0.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
indent-width = 2

[tool.ruff.format]
indent-style = "tab"
docstring-code-format = true

[tool.hatch.build.targets.wheel]
packages = ["redneck", "redplug"]

[tool.pyright]
typeCheckingMode = "basic"
